backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pydantic[email]==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
python-dateutil==2.8.2
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
playwright==1.40.0
slowapi==0.1.9